PyCuAmpcor: compile files as pure C++ when possible
This speeds up compilation, and brings it closer in line with the CPU port
