+++ contrib/PyCuAmpcor/src/cuAmpcorParameter.cpp
