+++ contrib/PyCuAmpcor/src/cuCorrNormalizer.cpp
