+++ contrib/PyCuAmpcor/src/cuOverSampler.cpp
