+++ contrib/PyCuAmpcor/src/cudaError.h
@@ -10,51 +10,13 @@
 
 #pragma once
 
-#include <cstdlib>
-#include <cstdio>
-#include <cstring>
-#include <cufft.h>
 #include "debug.h"
-#include <cuda.h>
-
-
-#ifdef __DRIVER_TYPES_H__
-#ifndef DEVICE_RESET
-#define DEVICE_RESET cudaDeviceReset();
-#endif
-#else
-#ifndef DEVICE_RESET
-#define DEVICE_RESET
-#endif
-#endif
 
 template<typename T >
-void check(T result, char const *const func, const char *const file, int const line)
-{
-    if (result)
-    {
-
-        fprintf(stderr, "CUDA error at %s:%d code=%d(%s) \n",
-                file, line, static_cast<unsigned int>(result), func);
-        DEVICE_RESET
-        // Make sure we call CUDA Device Reset before exiting
-        exit(EXIT_FAILURE);
-    }
-}
+void check(T result, char const *const func, const char *const file, int const line);
 
 // This will output the proper error string when calling cudaGetLastError
-inline void __getLastCudaError(const char *errorMessage, const char *file, const int line)
-{
-    cudaError_t err = cudaGetLastError();
-
-    if (cudaSuccess != err)
-    {
-        fprintf(stderr, "%s(%i) : CUDA error : %s : (%d) %s.\n",
-                file, line, errorMessage, (int)err, cudaGetErrorString(err));
-        DEVICE_RESET
-        exit(EXIT_FAILURE);
-    }
-}
+void __getLastCudaError(const char *errorMessage, const char *file, const int line);
 
 // This will output the proper CUDA error strings in the event that a CUDA host call returns an error
 #ifdef CUDA_ERROR_CHECK
